backend/requirements.txt
fastapi[standard]==0.115.0
pydantic-settings==2.6.0
google-genai==1.56.0
uvicorn==0.32.0